TESTING CI/CD

See merge request p2207446/mif-13!2

+++ users/src/main/java/fr/univlyon1/m1if/m1if13/users/service/UserResourceService.java
@@ -14,10 +14,11 @@ import javax.naming.NameAlreadyBoundException;
 import javax.naming.NameNotFoundException;
 import java.net.URI;
 import java.util.Collection;
+import java.util.NoSuchElementException;
 
 /**
  * Méthodes de service du contrôleur de ressources sur les utilisateurs.
-  */
+ */
 @Service
 public class UserResourceService {
 
@@ -36,26 +37,33 @@ public class UserResourceService {
                 .map(User::getLogin)
                 .map(s -> "users/" + s)
                 .map(LinkDto::new)
-                .toList()
-        );
+                .toList());
     }
 
-    public URI createUser(User user) throws NameAlreadyBoundException {
+    public URI createUser(User user) throws NameNotFoundException, NameAlreadyBoundException {
         userDao.add(user);
         return URI.create("users/" + user.getLogin());
     }
 
     public UserResponseDto getUser(String login) throws NameNotFoundException {
+        if (userDao.findOne(login) == null) {
+            throw new NoSuchElementException("Utilisateur non trouvé");
+        }
         return UserResponseDto.of(userDao.findOne(login));
     }
 
-    public void updateUser(String login, User user, String origin, HttpServletResponse response) {
+    public void updateUser(String login, User user, String origin, HttpServletResponse response)
+            throws NameNotFoundException {
+
         userDao.update(login, user);
         String token = userTokenProvider.generateToken(user, origin);
         response.addHeader("Authorization", "Bearer " + token);
     }
 
     public void deleteUser(String login) throws NameNotFoundException {
+        if (userDao.findOne(login) == null) {
+            throw new NoSuchElementException("Utilisateur non trouvé");
+        }
         userDao.deleteById(login);
     }
 }
